Media Breakpoints & Graph updates
The UI layout was reformatted slightly to make it easier for the dashboard to flex and respond properly to screen width. The dashboard converts (from a grid layout) to a single column at the 770px width (mobile devices) and/or a 600px height (small laptops).

Plotly graphs had y = 0 lines re-added as well as x-axis titles.